Updated factorial functions fact::Int->Double to fact::Double->Double

+++ cml/include/cml/math/fact.h
@@ -3,7 +3,7 @@
 
 __CML_BEGIN_DECLS
 
-double cml_fact(int x);
+double cml_fact(double x);
 
 __CML_END_DECLS
